requirements.txt
blessed>=1.20.0,<2.0.0
pygame>=2.5.2,<3.0.0
pydantic>=2.7.0,<3.0.0
pyyaml>=6.0.1,<7.0.0
rich>=13.7.0,<14.0.0
pytest>=8.2.0,<9.0.0
